backend/requirements.txt
fastapi
uvicorn
google-cloud-speech
google-generativeai
python-dotenv

  calisitmak icin 
  uvicorn main:app --reload --host 0.0.0.0 --port 8000

